Improve comparison of warn and crit levels
Currently it is not possible (for example) to set the check to "CRIT" when there is no consumer subscribed to a queue.
When defining "0" as crit or warn value the comparison returns "False" and does not set the state correctly. 
As far as I can tell from my testing "is not None" works fine here.